Store type info as pairs of instruction orders and types
Change AnnotationElement name from typeOfVreg to _TypeOfInstruction.

Add the pass TypeAdapter. This pass can convert the typeinfo of the form of the pairs of virtual registers and types to the form of pairs of instruction orders and types.

Add option --enable-typeinfo with which the pass TypeAdapter is enabled. Should be deleted when ark_js_runtime adaption is ready.

Add option --display-typeinfo with which the result typeinfo will be printed.

Issue: #I5E22F

Signed-off-by: qiuyu <qiuyu22@huawei.com>
Change-Id: I3c87b579d4c58cc3713263fe1e1e1d2ff953c58d